+++ CHANGES.rst
@@ -9,6 +9,9 @@ History
   Fixes `issue  #95 <https://github.com/collective/pas.plugins.ldap/issues/95>`_.
   [reinhardt]
 
+- Grant the Member role to all LDAP users.
+  [reinhardt]
+
 
 1.7.0 (2020-01-22)
 ------------------
@@ -29,7 +32,6 @@ History
 - Log long running LDAP/ pas.plugin.ldap operations as error.
   Threshold can be controlled with environment variable ``PAS_PLUGINS_LDAP_LONG_RUNNING_LOG_THRESHOLD``.
   [jensens]
-
 
 
 1.6.2 (2019-09-12)